feat: Make pending lint messages map building public
